Make bill moved into the bills fragment, and a huge logic had to be implemented, we gotta think this better.
Also there is a random crash that I haven't been able to reproduce well, it has to do with animating the charts, I try/catched it for now but there has to be something else.

Also moved the MakePaymentWithConekta, and I am passing the activity as a parameter, dont know the real impact of this, it appears to be minimal or none at all.